exercises 4.15 - 4.22

+++ part4/blog-list/controllers/blogs.js
@@ -1,8 +1,10 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
@@ -17,25 +19,42 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: user._id
   })
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  const user = await User.findById(decodedToken.id)
+  const blogToRemove = await Blog.findById(request.params.id)
+  if (user._id.toString() === blogToRemove.user.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+  } else {
+    response.json('unauthorized user')
+  }
   response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
   const body = request.body
-  console.log('body', body)
   const blog = {
     title: body.title,
     author: body.author,
@@ -54,7 +73,10 @@ blogsRouter.patch('/:id', async (request, response) => {
     return response.status(400).send('content missing').end()
   }
 
-  let blog = await Blog.findById(request.params.id)
+  response.json(
+    await Blog.findByIdAndUpdate(request.params.id, request.body, { new: true })
+  )
+  /*   let blog = await Blog.findById(request.params.id)
   blog = blog.toObject()
 
   Object.keys(blog).forEach(key => {
@@ -67,7 +89,7 @@ blogsRouter.patch('/:id', async (request, response) => {
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
     new: true,
   })
-  response.json(updatedBlog)
+  response.json(updatedBlog) */
 })
 
 module.exports = blogsRouter
